Fix line renderer
Updated the line renderer to use world space instead of the camera
space. Due to the game utilizing a canvas, the line renderer will never
be seen on screen when using the camera space. The renderer will now be
manaully updated with the world coordinates and drawn to the screen.