Implement updated camera logic
Camera now fully functions with new bounds. While between outside and
inside boudary, if the player move toward the inside boundary, the
camera moves as well, leading the player in the direction they move. If
the player attempts to move beyond the outside boundary, the camera
snaps ahead of the player, setting them insde the opposite set of
boundaries.